Support numbers 100 to 999
Add support for groups of blocks up to 999. New nodes are defined for
groups of 100 blocks, similar to 10s but with colours encoded in param2
using 3 separate palettes (4 groups per palette, since each group of 100
is checkered with light and dark).

Also add multipart number nodes support, including improved finding of
number nodes and determination of which set of blocks each belongs to
(using 2 upper bits of param2 to encode the direction the number is
extended from the central block).

Signed-off-by: James Hogan <james@albanarts.com>