Cleanup ten/hundred blocks
Currently ten and hundred blocks extend slightly beyond the block in
order to achieve the bordered effect. Unfortunately this doesn't work
well when neighbouring blocks are different colours as they flicker due
to depth buffering.

Therefore reduce the dimentions to fit within the block, and add fixed
and connected node boxes so that the insides aren't visible (and also
this should allow versions prior to 5.0.0 to show something for these
blocks).

Signed-off-by: James Hogan <james@albanarts.com>